+++ test/random.jl
@@ -5,8 +5,8 @@ using Main.TestHelpers.OAs
 
 using Base.Random: Sampler, SamplerRangeFast, SamplerRangeInt
 
-# Issue #6573
-guardsrand(0) do
+@testset "Issue #6573" begin
+    srand(0)
     rand()
     x = rand(384)
     @test find(x .== rand()) == []